refactor: extract useDebounced hook
