+++ src/hooks/useDebounced.js
@@ -1,0 +1,12 @@
+import { useEffect, useState } from "react";
+
+export default function useDebounced(value, delay = 300) {
+  const [v, setV] = useState(value);
+
+  useEffect(() => {
+    const id = setTimeout(() => setV(value), delay);
+    return () => clearTimeout(id);
+  }, [value, delay]);
+
+  return v;
+}
